Changed all instances of float to double

+++ Common/Game/Managers/Implementations/EventManager.cs
@@ -62,7 +62,7 @@ namespace Common.Game.Managers.Implementations
       return true;
     }
 
-    public void Update(float deltaTime, float maxTime)
+    public void Update(double deltaTime, double maxTime)
     {
       // queues are swapped so that any new events added in response to an 
       // event firing are processed in the next frame
